% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glyco-sa.R
\name{ageTMP_glyco_locfdr_settings}
\alias{ageTMP_glyco_locfdr_settings}
\title{Return manuscript glyco local-FDR settings}
\usage{
ageTMP_glyco_locfdr_settings(sex = c("Female", "Male"), adjusted = FALSE)
}
\arguments{
\item{sex}{One of \code{"Female"} or \code{"Male"}.}

\item{adjusted}{Whether the glyco result is protein-adjusted.}
}
\value{
A list of arguments suitable for \code{\link[=ageTMP_add_glyco_locfdr_columns]{ageTMP_add_glyco_locfdr_columns()}}.
}
\description{
Return manuscript glyco local-FDR settings
}
