% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpsa-protein.R
\name{ageTMP_load_cdisc_rna_matrix}
\alias{ageTMP_load_cdisc_rna_matrix}
\title{Load cDisc RNA matrix}
\usage{
ageTMP_load_cdisc_rna_matrix(data_dir = "data", collapse = TRUE)
}
\arguments{
\item{data_dir}{Path to the public data directory.}

\item{collapse}{Whether to average rows with the same feature identifier.}
}
\value{
A numeric gene-by-sample RNA matrix.
}
\description{
Load cDisc RNA matrix
}
