% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-public-data.R
\name{ageTMP_load_clinical}
\alias{ageTMP_load_clinical}
\title{Load public manuscript clinical data}
\usage{
ageTMP_load_clinical(data_dir = "data")
}
\arguments{
\item{data_dir}{Path to the public data directory.}
}
\value{
A data frame with clinical metadata.
}
\description{
Load public manuscript clinical data
}
